Improve design of the page
Done with help of Claude, by telling it first to
"improve the styling to make it look modern" and then
giving it a few specific prompts for further improvements.

It made some obvious mistakes which I fixed by manually
tweaking the CSS but overall I'm much happier with the look.